ProgLED and ProgButton pins are now optional
New init function for own ProgLED and ProgButton implementation
New DemoSketch for new init-function
DemoSketches updatet to beta4a

+++ examples/DemoSensor_DS18B20/kdevice_Temperature_DS18B20.h
@@ -31,4 +31,4 @@ byte KonnektingDevice::_paramSizeList[] = {
     /* Index 6 - tempMaxValue */ PARAM_UINT8,
     /* Index 7 - tempMaxLimit */ PARAM_INT16
 };
-const byte KonnektingDevice::_numberOfParams = sizeof (_paramSizeList); // do not change this code
+const int KonnektingDevice::_numberOfParams = sizeof (_paramSizeList); // do not change this code
